Fixed DocRef equals and hashcode methods and made some minor code improvements as a result of analysis. Updated version to v0.1.1.

+++ stroom-query-common/src/test/java/stroom/query/TestChartResultCreator.java
@@ -82,6 +82,6 @@ public class TestChartResultCreator {
     }
 
     private double getYValue() {
-        return Double.valueOf(Math.random() * 1000D);
+        return Math.random() * 1000D;
     }
 }
